weatherWidget refactoring
